requirements.txt
asgiref
autopep8
psycopg2
django
django-filter
django-environ
django-rosetta
django-widget-tweaks
django-extensions
dj-database-url

djangorestframework
djangorestframework-datatables

drf-spectacular

pycodestyle
pytz
sqlparse
Unipath
python-decouple
gunicorn
whitenoise
